Update windows-failed-to-apply-mdm-policy.md
I had a device where it was supposed to be enrolled in Intune via GPO. There was a registry key in Computer\HKEY_LOCAL_MACHINE\SOFTWARE\Microsoft\Enrollments\ that showed it should have been enrolled but Intune never saw the device. Because of this, the device would refuse the MDM enrollment since it was "already enrolled". After deleting this key and running GPUPDATE the device enrolled successfully.